Handle messages retrieval failure
When the `IStorageConnection` service is down, `connection.GetPublishedMessagesOfNeedRetry()` and `connection.GetReceivedMessagesOfNeedRetry()` both throw an exception.

The wait operation in `ProcessAsync` is therefore never executed.

We therefore end up in an infinite loop of immediate retries until the `IStorageConnection` service is back up. This can be seen in the logs where we have hundreds of "Processor 'DotNetCore.CAP.Processor.NeedRetryMessageProcessor' failed. Retrying..." entries per second.

We add a try/catch around the message retrieval operation and log the error. The next messages retrieval will therefore be executed `options.FailedRetryInterval` seconds later.