Fixes for OpenTelemetry Context Propagation
1. Fixes an issue that caused context propagation to fail on message
retry (#1452)
Before, BeforeSubscribeInvoke removed the context so it was no longer
available on retry.

It also removes the _contexts dictionary, as during Subscribe this
handling didn't work during retry and during Publish, the dictionary
value was no longer used.

I have tested this with both aspnetcore services and console subscriber
that publishes during the subscribe and context propagation was working
fine. Thus #1100 is also closed, probably already in #1407.